Use ecore.Exe instead of multiprocessing

+++ π.py
@@ -3,75 +3,20 @@
 
 from __future__ import absolute_import, print_function, unicode_literals
 
+import os
 
-##############################################################################
-# Unbounded Spigot Algorithms for the Digits of Pi
-# http://web.comlab.ox.ac.uk/oucl/work/jeremy.gibbons/publications/spigot.pdf
-# S. Rabinowitz and S. Wagon
-
-# http://davidbau.com/archives/2010/03/14/python_pipy_spigot.html
-def pi_decimal_generator():
-  q, r, t, j = 1, 180, 60, 2
-  while True:
-    u, y = 3*(3*j+1)*(3*j+2), (q*(27*j-12)+5*r)//(5*t)
-    yield str(y)
-    q, r, t, j = 10*q*j*(2*j-1), 10*u*(q*(5*j-2)+r-y*t), t*u, j+1
-
-## Usage:
-# count = 0
-# digits = pi_decimal_generator()
-# while 1:
-#   line = ''.join([digits.next() for j in xrange(50)])
-#   print '%6d: %s' % (count, line)
-#   count += 50
-
-##############################################################################
-
-from multiprocessing import Process, Queue, Event
-
-from efl import ecore
 from efl import evas
-from efl.evas import EVAS_HINT_EXPAND, EVAS_HINT_FILL, \
-    EVAS_TEXTGRID_PALETTE_STANDARD
-
+from efl import ecore
 from efl import elementary
 from efl.elementary.window import StandardWindow
 from efl.elementary.label import Label
 from efl.elementary.entry import Entry
 
-EXPAND_BOTH = EVAS_HINT_EXPAND, EVAS_HINT_EXPAND
-EXPAND_HORIZ = EVAS_HINT_EXPAND, 0.0
-EXPAND_VERT = 0.0, EVAS_HINT_EXPAND
-FILL_BOTH = EVAS_HINT_FILL, EVAS_HINT_FILL
-FILL_HORIZ = EVAS_HINT_FILL, 0.5
-FILL_VERT = 0.5, EVAS_HINT_FILL
+from efl.evas import EVAS_HINT_EXPAND, EVAS_HINT_FILL
 
 
 COLS = 50
 ROWS = 15
-
-
-class PigrecoThread(Process):
-    def __init__(self, queue):
-        Process.__init__(self)
-
-        self._queue = queue
-        self._stop = Event()
-        self._generator = pi_decimal_generator()
-
-    def stop(self):
-        self._stop.set()
-
-    def run(self):
-        next(self._generator)
-        while 1:
-            line = ''.join([next(self._generator) for j in range(COLS)])
-            self._queue.put(line)
-
-            if self._stop.is_set():
-                break
-
-        print("DONE")
 
 
 class PiWin(StandardWindow):
@@ -99,8 +44,8 @@ class TG(evas.Textgrid):
                                size=(COLS, ROWS),
                                font=("Monospace", 15))
 
-        self.palette_set(EVAS_TEXTGRID_PALETTE_STANDARD, 0, 0, 0, 0, 100)
-        self.palette_set(EVAS_TEXTGRID_PALETTE_STANDARD, 1, 150, 150, 150, 255)
+        self.palette_set(evas.EVAS_TEXTGRID_PALETTE_STANDARD, 0, 0, 0, 0, 100)
+        self.palette_set(evas.EVAS_TEXTGRID_PALETTE_STANDARD, 1, 150, 150, 150, 255)
 
         pxw, pxh = self.cell_size
         self.size_hint_min = (pxw * COLS, pxh * ROWS)
@@ -114,71 +59,83 @@ class TG(evas.Textgrid):
             self.cellrow_set(row_count, row)
 
         self.lines = lines
+        self.scroll_pos = 0
 
-    def redraw(self, beg_line=None):
-        if beg_line is None:
-            beg_line = max(0, len(self.lines) - ROWS)
+    def scroll_to(self, line):
+        self.scroll_pos = min(line, len(self.lines) - ROWS)
 
+    def scroll_rel(self, offset):
+        self.scroll_to(self.scroll_pos + offset)
+
+    def redraw(self):
+        # TODO avoid redraw if we are still at the bottom
         for row_num in range(ROWS):
             if row_num >= len(self.lines):
                 break
+
             row = self.cellrow_get(row_num)
             for i, cell in enumerate(row):
-                cell.codepoint = self.lines[beg_line + row_num][i]
+                cell.codepoint = self.lines[self.scroll_pos + row_num][i]
             self.cellrow_set(row_num, row)
 
         self.update_add(0, 0, COLS, ROWS)
 
-    def line_append(self, line):
-        self.lines.append(line)
-        self.redraw()
 
 
 class Pigreco(object):
     def __init__(self):
         self.lines = []
         self.count = 0
-        self.queue = Queue()
+        self.exe = None
 
         self.win = PiWin(self)
-        
-        self.thread = PigrecoThread(self.queue)
-        self.thread.start()
 
+        for i in range(ROWS-1):
+            self.lines.append(' ' * COLS)
         self.lines.append('3.' + ' ' * (COLS-2))
         self.win.tg.redraw()
 
         self.timer = ecore.Timer(1.0, self._timer_cb)
+        self.generator_start('generator_spigot.py')
+
+    def generator_start(self, executable):
+        self.exe = ecore.Exe('python %s %d' % (executable, COLS),
+                             ecore.ECORE_EXE_PIPE_READ |
+                             ecore.ECORE_EXE_PIPE_READ_LINE_BUFFERED)
+        self.exe.on_data_event_add(self._generator_stdout)
+        # exe.on_del_event_add(self._cmd_done)
+
+    def _generator_stdout(self, exe, event):
+        self.lines.extend(event.lines)
+        self.count += len(event.lines) * COLS
+        self.win.title = 'π  - {:,} decimals'.format(self.count)
 
     def _timer_cb(self):
-        if self.queue.empty():
-            return ecore.ECORE_CALLBACK_RENEW
-
         if self.timer.interval > 0.1:
             self.timer.interval -= 0.02
 
-        line = self.queue.get()
-        self.count += COLS
-        print(self.queue.qsize(), self.count, line)
-
-        self.lines.append(line)
+        self.win.tg.scroll_rel(+1)
         self.win.tg.redraw()
-        self.win.title = 'π  (decimals: %d)' % self.count
 
         return ecore.ECORE_CALLBACK_RENEW
 
     def quit(self):
-        self.timer.delete()
-        self.thread.stop()
-        self.thread.join()
         elementary.exit()
+
+    def cleanup(self):
+        if self.timer:
+            self.timer.delete()
+        if self.exe:
+            self.exe.terminate()
+        
 
 
 if __name__ == "__main__":
     elementary.init()
 
-    Pigreco()
-
+    app = Pigreco()
     elementary.run()
+
+    app.cleanup()
     elementary.shutdown()
